requirements.txt
python-dotenv==1.0.0
pytube
spotipy==2.23.0
moviepy